+++ src/util/console.js
@@ -1,0 +1,9 @@
+function getConsole() {
+  if (typeof window !== "undefined") {
+    return window.console;
+  }
+  return global.console;
+}
+const console = getConsole();
+
+export { console };
