+++ tests/unit/util/console.spec.js
@@ -1,4 +1,4 @@
-import { console } from "@/util/helper";
+import { console } from "@/util/console";
 
 describe("console", () => {
   test.each([
